Documentation
This is the project of a node.js server useful to implement an instant messaging web chat. You che use it for all common task of a chat, like send message, view online users, send when other users are writing to you and so on... 

Feel free to use it in your projects. But please notice and refer me.